+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed columnPicker to prevent the mutation of the props
 - Fixed the no uniqueness of widgets ids by replacing them by class
 - Fixed checkbox props propagation
+- Fixed the aggregation step when performing an aggregation on one of the id columns
 
 ## [0.15.0] - 2020-03-24