+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED -
+## [0.15.1] - 2020-04-06
 
 ### Fixed
 
@@ -217,6 +217,7 @@
 
 - Initial version, showtime!
 
+[0.15.1]: https://github.com/ToucanToco/weaverbird/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/ToucanToco/weaverbird/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/ToucanToco/weaverbird/compare/v0.13.1...v0.14.0
 [0.13.1]: https://github.com/ToucanToco/weaverbird/compare/v0.12.0...v0.13.1